requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
flask-limiter>=3.5.0
pymongo>=4.0.0
nltk>=3.8.0
bcrypt>=4.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
gunicorn>=21.0.0
python-docx>=1.1.0
dnspython>=2.4.0
